% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{KM31}
\alias{KM31}
\title{Krishnamoorthy & Mathew's example 3.1}
\format{A data frame with 16 rows and 3 numerical columns.}
\usage{
data(KR31)
}
\description{
The dataset used in Krishnammorthy & Mathew's example 3.1.
}
\examples{
data(KM31)
str(KM31)
}
\references{
Krishnamoorthy and Mathew, Statistical Tolerance Regions, Wiley 2009.
}
\keyword{data}
