% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolfactor}
\alias{tolfactor}
\title{Tolerance factor for linear regression}
\usage{
tolfactor(X, xnew, p = 0.9, alpha = 0.05, side = "one-sided", ...)
}
\arguments{
\item{X}{model matrix}

\item{xnew}{new values of the predictors}

\item{p}{proportion to be covered by the tolerance interval}

\item{alpha}{significance level}

\item{side}{either \code{"one-sided"} or \code{"two-sided"}}

\item{...}{arguments passed to \code{\link[base]{uniroot}} if \code{side="two-sided"}, 
otherwise ignored}
}
\value{
A number, the tolerance factor. If \code{side="two-sided"}, some attributes 
are set, the components of the output of \code{\link[base]{uniroot}}.
}
\description{
Tolerance factor for linear regression
}
\examples{
fit <- lm(mpg ~ disp+drat, data = mtcars)
tolfactor(model.matrix(fit), xnew=c(1, 160, 4), side="two-sided") 
data(KM31)
fit <- lm(y ~ x1+x2, data = KM31)
tolfactor(model.matrix(fit), xnew=c(1, 88, 9), side="one-sided")
tolfactor(model.matrix(fit), xnew=c(1, 88, 9), side="two-sided")$root
}
